Start rewriting to use SQL
This commit is a major jump toward a SQL-based system.  It eliminates
packages from the system, opting instead to store all data in
$XDG_DATA_HOME/ppaml and indexing said data using a SQLite table.  The
'ppaml eval' command has not yet been ported to the new world order of
SQLite, so I’ve removed it for the time being.

Everything seems to work correctly, but it is likely very brittle.
Better exception handling is still a TODO.

The documentation has not yet been updated.