Don't take the character-wise common prefix of directory.
It seems like a bad idea and bit me in a funny way. I had a directory
with two subdirectories in it - testing/ and training/. They both
start with a 't', thus we ended up getting `<dir>/t` instead of
`<dir>`. I don't understand why we would want the directory prefix,
so maybe this will come back to haunt me. Oh well.