Unpack dataset input directory and pass to evaluator
When evaluating a run, `peval` now unpacks the input directory of
the runs dataset and passes it to `eval.sh` as an environment
variable: "INPUT_DIR". This allows the evaluator to reuse input
data, rather than having to duplicate it into the "eval" directory.

We pass the directory path with an environement variable rather than
a command argument to maintain backwards compatibility (old evaluators
would have their positional arguments messed up).